+++ app/src/components/SearchBar/index.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {Image, TextInput} from 'react-native';
+import {ActivityIndicator, Image, TextInput} from 'react-native';
 import ElevatedView from 'react-native-elevated-view';
 import textStyles from '../../theme/parkspotStyles';
 import styles from './styles';
@@ -27,17 +27,22 @@ export default class SearchBar extends Component {
           source={require('../../../assets/icons/misc/search.png')}
         />
         <TextInput
-          placeholder="Search for a parkspot"
+          placeholder='Search for a parkspot'
           returnKeyType={'search'}
           autoFocus={true}
           value={this.props.searchString}
-          clearButtonMode="while-editing"
-          underlineColorAndroid="transparent"
+          clearButtonMode='while-editing'
+          underlineColorAndroid='transparent'
           onChangeText={text => this.props.onChange(text)}
           onFocus={() => {
             this.props.onFocus ? this.props.onFocus() : null;
           }}
           style={[styles.input, textStyles.textStyle2]}
+        />
+        <ActivityIndicator
+          animating={true}
+          size='small'
+          style={{marginRight: 10, display: this.props.isLoading ? 'flex' : 'none'}}
         />
       </ElevatedView>
     );