+++ app/src/container/MapContainer/reducer.js
@@ -23,16 +23,16 @@ export default function (state: any = initialState, action: Function) {
       }
     };
   } else if (action.type === 'FETCH_PARKSPOTS_SUCCESS') {
-    var combined;
+    let combined;
     if (!action.refresh) {
       // adding all new parkspots to the data
       combined = _.unionBy(state.parkspots, action.data, 'id')
       // updating the values for all existing parkspots without recreating
       // the object to prevent the map from rerendering too much.
-      for (var old in state.parkspots) {
-        var updated = action.data.find(p => p.id === old.id);
+      for (let old in state.parkspots) {
+        const updated = action.data.find(p => p.id === old.id);
         if (updated) {
-          for (var key in old) {
+          for (let key in old) {
             old[key] = updated[key];
           }
         }
@@ -41,6 +41,12 @@ export default function (state: any = initialState, action: Function) {
       combined = action.data;
     }
 
+    // mocking some data TODO use API when ready
+    state.parkspots.map(parkspot => {
+      parkspot.unlimited = Math.random() < 0.5;
+      parkspot.noCost = Math.random() < 0.5;
+    });
+
     // Apply current filters on fetched parkspots
     let filteredParkspots = [];
     filteredParkspots = combined.filter(obj => {
@@ -48,7 +54,7 @@ export default function (state: any = initialState, action: Function) {
       state.filters.forEach(filter => {
         if (!obj[filter]) {
           showParkspot = false;
-          return;
+
         }
       });
       return showParkspot;
@@ -74,7 +80,6 @@ export default function (state: any = initialState, action: Function) {
         filters.forEach(filter => {
           if (!obj[filter]) {
             showData = false;
-            return;
           }
         });
         return showData;
@@ -83,10 +88,7 @@ export default function (state: any = initialState, action: Function) {
       // Filter turned on -> adds filter on top of filteredParkspots
       filters.push(action.filter);
       filteredParkspots = state.parkspots.filter(obj => {
-        if (!obj[action.filter]) {
-          return false;
-        }
-        return true;
+        return obj[action.filter];
       });
     }
     return {