+++ app/src/container/SearchContainer/index.js
@@ -82,7 +82,7 @@ const mapDispatchToProps = dispatch => {
       dispatch(remFavorite(fav));
     },
     addLastSearched: place => {
-      dispatch(addLastSearched(place))
+      dispatch(addLastSearched(place));
     }
   };
 };