+++ app/src/theme/parkspotColors.js
@@ -4,6 +4,9 @@ export default {
   lightAquamarine: '#8affc1',
   steel: '#8e8e93',
   blackTwo: '#1d1e18',
+  blackThree: '#030303',
   cement: '#989896',
   greyishTeal: '#7a9e9f',
+  gray: 'rgba(152, 152, 150, 0.55)',
+  placeHolderNative: '#C7C7CD',
 };